Final edits for editing a course
I can only assume that this works, as I cannot for the life of me get any testing progress to be made. Ideally someone else will see this and know exactly what needs to changing.

Made edits relating to passing data to the api for updating the course we're trying to edit